+++ pkg/version/version_test.go
@@ -258,11 +258,11 @@ func TestCurrentCompatibilityManifestUsesOnePlatformVersion(t *testing.T) {
 func TestCurrentPackageSetFallsBackToBuildInfoVersion(t *testing.T) {
 	restore := replaceReleaseVersions(devVersion, devVersion, devVersion)
 	defer restore()
-	restoreDetector := replaceBuildInfoDetector("0.7.0")
+	restoreDetector := replaceBuildInfoDetector("0.7.1")
 	defer restoreDetector()
 
 	packageSet := CurrentCompatibilityManifest().PackageSet
-	if packageSet.PlatformVersion != "0.7.0" || packageSet.ContractSetSHA256 != ContractSetSHA256 {
+	if packageSet.PlatformVersion != "0.7.1" || packageSet.ContractSetSHA256 != ContractSetSHA256 {
 		t.Fatalf("package set = %#v", packageSet)
 	}
 }